Safari-611.1.10/LayoutTests/fast/dom/window-languagechange-event-expected.txt

Tests the languagechange event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.language is "en-US"
* received languagechange event
PASS navigator.language is "fr-FR"
PASS successfullyParsed is true

TEST COMPLETE
